refactor all of the permalink generation functions into a new set of modules organized by source (i.e. GPO, Library of Congress, etc.)
fixes #131
