+++ citations/dc_law.js
@@ -29,22 +29,5 @@ module.exports = {
         }
       }
     ];
-  },
-
-  links: function(cite) {
-    // Only available for CP's 19 and 20.
-    if (cite.period < 19 || cite.period > 20) return { };
-    return {
-      dccodeorg: {
-        source: {
-            name: "Council of the District of Columbia",
-            abbreviation: "DC Council",
-            link: "https://dccode.gov",
-            authoritative: true
-        },
-
-        pdf: "https://dcstat.dccode.gov/public/Law_" + cite.period + "-" + cite.number + ".pdf"
-      }
-    };
   }
 };
