+++ citations/law.js
@@ -71,33 +71,5 @@ module.exports = {
         };
       }
     }
-  ],
-
-  links: function(cite) {
-    var links = { };
-    if (cite.congress >= 104) {
-      links.usgpo = {
-        source: {
-          name: "U.S. Government Publishing Office",
-          abbreviation: "US GPO",
-          link: "http://www.gpo.gov",
-          authoritative: true
-        },
-        pdf: "http://api.fdsys.gov/link?collection=plaw&congress=" + cite.congress + "&lawtype=" + cite.type + "&lawnum=" + cite.number,
-        mods: "http://api.fdsys.gov/link?collection=plaw&congress=" + cite.congress + "&lawtype=" + cite.type + "&lawnum=" + cite.number + "&link-type=mods"
-      };
-    }
-    if (cite.congress >= 82) {
-      links.govtrack = {
-        source: {
-          name: "GovTrack.us",
-          abbreviation: "GovTrack.us",
-          link: "https://www.govtrack.us",
-          authoritative: false
-        },
-        landing: "https://www.govtrack.us/search?q=" + (cite.type=="public"?"Pub":"Priv") + "Law+" + cite.congress + "-" + cite.number
-      };
-    }
-    return links;
-  }
+  ]
 };
